Fix issues with qwidgetwidget and qwidgettexture
The qwidgettexture class mixes Qt opengl with VTK opengl which can make
state management tricky. The prior code tried to share a texture handle
between the two which caused OpenGL errors in some cases. This MR takes
a safer approach and renders the Qt content to an image and then takes
the image data and creates a txture from it. This may be a bit slower
but reduces the OpenGL interactions between the two systems.

There is an issue where in VR shaky hands can lead to drag selections
when clicks were intended. This MR includes a timer to make clicks
that are one second or less in duration have the same position
for the mouse events (preventing it being interpreted as a drag)