vtkPlaneCutter: CopyAllocate the correct CellData
When contour is used inside the operators of the functors,
since internally it takes the values from the inputCellData,
the cellTypeCellData need to also CopyAllocate the inputCellData.
This is necessary because the TargetIndices map of vtkDataSetAttributes
needs to be correct, otherwise, out of bounds errors can occur.

(cherry picked from commit 75d5b161447e32844fc0ea9e066c4dfa760169b1)