For pyi files, only include our own namespaces
We wrap C++ namespaces as modules, or more specifically as
instances of sub-type of module that we call "namespace".
When putting our stuff in the .pyi files, we want to include
our namespaces, but not other Python modules.  For example,
if someone does "import os" in __init__.py, we definitely
don't want to put "os" and all its contents in the .pyi.