vtkMultiObjectMassProperties: Compute centroid of objects
The main addition of this commit is the computation of the centroid
of each object as a weighted average of the centroids of the tetrahedrons
which are used to compute the volume of the polygonal object, and the weight is the
tetrahedron's volume contribution. This done in a multithreaded way.
Additionally, the computation of the object's volume and area is also done
a multithreaded way, which may lead to slightly different results (order of e^-17)
in each execution, because the order that doubles are added matters,
but yields greater performance.