Fixed issues with parsing for the H5Rage Reader
Fixed an issue where variable names were assumed to be only three
characters, but they actually can be longer.

Added a new option, called HDF_CYCLE_DIGITS, which specifies how many
digits are used for the cycle number at the end of the filename. Before
it was assumed it was always six digits, but actually it can be
different. If the option is not specified, it is assumed to be six
digits.