dataset_adapter.py: take array length from row data only if column exists
If a new wrapped vtkTable was being populated with append() calls, the
array length check added in 224d3db6 would return 0, and subsequent
logic would treat tuples in the array as components instead. Fix that
by only considering the length of row data if an existing column is
present already.