Add VR CameraPose to vtkVRCamera
Add a VR camera pose that captures VR specific properties and provides
methods to save/apply such poses.

To reproduce a pose using the current camera and render window means to
try to make the viewers current view look like the original saved pose.
This is complicated by the fact that the viewer may now occupy a very
different position and orientation in the physical space than when the
pose was saved. This method accounts for this and adjusts the physical
space to best fit the requested pose.