Reducing vtkGhostCellsGenerator compile-time.
This commit creates methods in `vtkDIYGhostUtilities` that uses the
correct templated version of `vtkDIYGhostUtilities::GenerateGhostCells`.
This allows the entire pipelines to be compiled when `vtkDIYGhostUtilities` is compiled.
On my machine, the compile-time of `vtkGhostCellsGenerator` is reduced
by 3 seconds when applying this change (assuming that module `Parallel/DIY` is already compiled).