Fix(vtkDebugLeaks): fix double-mutex locking
Fixes issue #18362.
vtkDebugLeaks locks and unlocks the std::mutex* CriticalSection
explicitly. It also reports error and warning messages before unlocking
the mutex. If a failure occurs, the function tries
to report the error using vtkGenericWarningMacro. This macro might
construct a vtkWin32OutputWindow object which can lead to a second
attempt at locking the mutex.

Summary of Changes:
Added the user of std::lock_guard instead of explicit lock()/unlock()
calls. This guarantees that mutex will be unlocked when scope runs
out. Also, the scope of locking has been changed so that the mutex
is unlocked before printing / reporting any errors.