vtkParseMerge_Merge: fix indirect leak
`vtkParseMerge_Merge` invalidates a function of `super` (`super->Functions[ii] = NULL`)
no matter if there is a matching function in the derived class or not.
However, only the second case transfers ownership through
`vtkParse_AddFunctionToClass(merge, f1)` and as a result `f1` leaks when there
is a match (first case).

This commit avoids this issue by leaving `f1` valid in the fist case which
allows to free `f1` properly when `super` is freed.

This change affects only the `super` class, the result of `merge` remains
the same.