Fix Xdmf polyhedron testing file.
Fix a bug with the Xdmf polyhedron data file used for testing
the arbitrary polyhedra IO capabilities of Xdmf.
The dimensions for the geometry array were incorrectly specified.
While it did not cause an issue with Xdmf3, it is not correct
and does not work correctly with the Xdmf2 reader added in #17621.

See #17621.