vtkObjectBase: support a "debug" class name accessor
This is used for `vtkDebugLeaks` to be able to track `vtkCommand` and
its subclasses. This is different than normal `vtkObject` classes
because `vtkCommand` and its subclasses actually use public constructors
which do not use the `New()` method and as such, the classname is not
known at the time of registration which then mismatches with the
classname as known at destruction time.

Also fixes a typo in a docstring.