vtkRedistributeDataSetFilter: number of outputs bug fix
In `vtkRedistributeDataSetFilter`, when the input had a structure
where some rank had a multi block with data in it, and other blocks had
empty multi blocks, then the output in these ranks were squeezed out.

This was due to the `PreservePartitionsInOutput` that was forcing the
output and input to strictly share the same structure. Loosening it by
just enforcing cases where the output has less blocks than the input
fixes the issue