vtkXRenderWindowInteractor: Fix keyCode value
VTK follows the convention that keyCode and keySym number are the same.
When ctrl + letter is pressed, buffer[0] was returning the control ascii character,
instead of the keySym. This commit addresses this issue.

This convention is not correct though, and a future fix that removes the
keyCode needs to be implemented.