vtkRedistributeDataSet bug fix: remove duplicate points
`vtkRedistributeDataSet` could output duplicate points at the interfaces
where the previous partition splitting was occurring. This duplicate
could make downstream filters fail (such as `vtkGhostCellsGenerator`).
This could be worked around by calling a cleaning filter right after
`vtkRedistributeDataSet`.

This commit makes the cleaning step unnecessary. The point duplication
was caused by calls to `vtkAppendFilter` on which `MergePoints` was
turned off.

A call to `vtkAppendFilter` at the very end of the pipeline is also removed
in the instance of `vtkUntructuredGrid` output.

Adresses paraview/paraview#21161