Added point iterator
This helper class is used to iterate over sets of points in a
variety of ways using the concept of a spherical region of
iteration. Depending on the axes provided (which emanate from
a center of rotation, and whether the points are radially
sorted), points surrounding a center point can be iterated over
in a local neighborhood.