Add a vtkPythonUtil::IsInitialized() check
The Python wrappers require vtkPythonUtil::Initialize() to be
called before the wrappers are used.  This is done during the
module initialization for vtkCommonCore.  When other modules
are loaded, they call vtkPythonUtil::IsInitialized(), which will
fail if their vtkPythonUtil is not the same one initialized by
vtkCommonCore.  For example, if vtkCommonCore came from a
pip installation of VTK, and you load another module that links
to a VTK installed somewhere else, an initialization failure will
occur.  Prior to this commit, the initialization falure caused a
segfault, but now it raises an ImportError exception instead.