vtkPIOReader: fix segfault in PIOAdaptor
PIOAdaptor.cxx had several global pointers that are
allocated/deallocated in the constructor/destructor of PIOAdaptor. This
led to memory leaks as well as the potential for double frees or trying
to use a pointer that had already been freed, when more than one
instance is created. These variables (along with other variables used to
save state) were moved to an Impl class within PIOAdaptor.