Fixed Bugs About Computing 3D Surface
A bug was caused by not using SetInputArrayToProcess in the
AdvectSimple function in the vtkStreamSurface filter. The other bug was caused by
the incorrect threshold usded in the AdvectIterative function. We
added a new test for testing the correctness of the AdvectSimple function.
Finally, we replaced the threshold for computing critical points with
EpsilonCriticalPoint, which can be changed by users.