vtkDIYGhostUtilities: loosen bounding boxes
If the input data sets have loose point positions (i.e. points across partitions
that are supposed to match don't actually match), the ghost cells generator
fails at binding those points. This is not a problem if the input does
not have global ids. However, if it has global ids, points then match.
If a partition A has a bounding box that doesn't include some neighbor's
points while this neighbor's bounding box includes the ones of A, then,
if some of those point's global ids match, there is a conflict on the
topology between partitions, and, as a consequence, the process will fail.
The issue is the lack of symmetry between blocks while running the
algorithm.

To avoid such events, each bounding box of each partition is now
inflated by a lot (0.1% of its max width). This will
include more points when checking which points are to be transfered,
but this will include any "decent" situation. From now on, the filter will
fail for points whose precision is lower than 0.1% of the bounding box widh
with global ids attached to the points. Such an input can be deemed
ill-formed.

Adresses paraview/paraview#21228

(cherry picked from commit 9a71560767ef0ca240138c43708eb4a4f4482498)