vtkGhostCellsGenerator: allowing to request 0 layers of ghost cells
The number of layers of ghosts cells to generate was clamped to [1,
VTK_INT_MAX]. This commit updates the clamping to allow 0 levels of
ghosts.

There were a few issues to address for structure data sets when
generating 0 levels of ghosts. They are addressed.

In addition, when requesting zero layers of ghost cells, when possible,
the output is shallow-copyied from the
input. This is done when the input doesn't have any ghost layers
already.

Ghost cells are not present in the output if requesting zero layers of
ghost cells and if the data set doesn't have hidden ghosts while being a
structured data set. It is necessary to preserve them for hidden ghosts
in structured data sets to preserve the input's geometry / topology.