Use absolute path for LICENSE_FILES property
The LICENSE_FILES listed in the vtk.module files should be assumed
to be relative to the path of the vtk.module file.  Previously, it
was assumed that they were relative to the path of the module
directory.  This fix is needed when the module file is not within
the module directory.