vtkDIYUtilities with non-standard memory layout array handling
The failing test was because vtkDIYUtilities does not report failure of
array dispatch, in casu when VTK_DISPATCH_SOA_ARRAYS=OFF (the default).
Support for non-standard memory layout arrays has been added, but is not
optimal in case of e.g. SOA arrays, i.e. a deep-copy is made prior to
serialization.