vtkPlaneCutter: Fix AMR & MultiBlockDataSet
The number of pieces of a vtkMultiPieceDataSet is now guaranteed
to be correct.

The output of each vtkDataSet is now a vtkPolyData instead of a
vtkMultiPieceDataSet which is produced by appending the pieces of
vtkMultiPieceDataSet using vtkAppendPolyData. This was done because
for a complex vtkMultiBlockDataSet in distributed fashion, CopyStructure fails.

The AMR version before this MR did not even work because vtkStructuredGrid
was not supported and GenerateCompositeDataSetFromHierarchy would generate
a vtkUniformGridAMR instead of a vtkMultiPieceDataSet, which based on the
output type that's wrong. For that reason now the output type is
vtkPartitionedDataSetCollection so that we can just shallow copy the result.

The old MultiBlockDataSet version with vtkDataAssemblyUtilities has
been replaced because vtkDataAssemblyUtilities would complain that
`More than 1 partition encountered. Only 1st will be copied over`.