Fix duplicate function definition when hid_t is int32.
In older versions of HDF5, the hid_t type can be a 32-bits integer; newer versions use a 64-bit integer.
When VTK is built to link with such an older version, the NewArray(...) function is defined twice in vtkHDFReaderImplementation.h with the same signature. This MR fixes that problem, by renaming one of the clashing functions.