vtkDataArray::GetRange becomes ghost aware
Methods `vtkDataArray::GetRange` and `vtkDataArray::GetFiniteRange` are
now ghost aware. A new version of those methods takes a `const unsigned char* ghosts`
and an `unsigned char ghostsToSkip` parameters. Any element from
`ghosts` that matches `ghostsToSkip` (a binary AND is performed to
assess it) are skipped when computing the range.

The parameter had to be a `const unsigned char*` rather than a
`vtkUnsignedCharArray*` because this data array type would need to be fully
defined when reaching `vtkGenericDataArray.txx`, which is not possible, as it is
derived from it.