Introducing vtkFieldData::GetRange
This new method leverages the new ghost aware `vtkDataArray::GetRange` API.
It uses the ghost array associated with the field data if available,
and a `GhostsToSkip` can be set in `vtkFieldData`. When adding a ghost
array, the new attribute `vtkFieldData::GhostArray` is set. When
removing the ghost array, this attribute is set to `nullptr`.

In particular, `vtkCellData::GhostsToSkip` is set to
`vtkDataSetAttributes::HIDDENCELL`, and `vtkPointData::GhostsToSkip` is
set to `vtkDataSetAttributes::HIDDENPOINT` by default.