Ghost cells generator: ghost point tagging bug fix for unstructured data
Some ghost points were weeded out if they weren't part of our duplicate
point book keeping. This was wrong because we do not necessarily receive
each ghost point twice. This is corrected in this commit.

The test is also updated to also test one layer of ghosts.

Adresses paraview/paraview#21348