Threaded implementation of the 3D Surface Nets isocontouring algorithm
Surface Nets is an algorithm used to isocontour segmented 3D volumes.
The segmentation is represented by a label map which classified each
voxel as inside a segmented region, or as part of the background.
One or more labeled regions can be extracted simultaneously. The
algorithm also includes a constrained smoothing process to produce
higher-quality results. The algorithm is threaded using concepts
from Flying Edges (which processes data independently on an edge-by-edge
basis).