vtkAbstractCellLocator: Ensure correctness for FindCell
Before a FindCell query, call IsInBounds to ensure correctness.

Also, for vtkStaticCellLocator use IsInBounds instead of
vtkMath::PointIsWithinBounds. The reason behind this non-easily
seen outcome is that PointIsWithinBounds validates that the given
parameters are not null before actually calculating the result that we want.

With this change on a specific streamlines experiment
PointIsWithinBounds took 1.810 sec, IsInBounds took 0.860 sec.
The speedup is 1.810 sec / 0.860 sec = 2.15 times better.