vtkThreshold: dumping input HIDDENCELL ghosts
`vtkThreshold` used to copy input `HIDDENCELL` that can be present in
structured data set types (`vtkImageData` for instance). Those cells do
not need to be copied to the output, and copying them is actually a
waste of memory and can lead downstream algorithm not expecting hidden
cells to be faulty. Hidden ghost cells are now dumped.