vtkAbstractInterpolatedVelocityField: Optimize FindAndUpdateCell
Previously FindCell would be called with the last cell and if
that would fail it would try to do the same without the initial cell.
This was WRONG because FindCell does that already(first checks the initial
cell, if that does not help, it uses the point locator) This way FindCell
cost is cut almost in half.

Calls to EvaluatePosition and GetCell has been reduced after careful
understanding of when they are needed (see comments).

The quality of the results has been improved when surfaceDataSet is on.

vtkCellLocatorStrategy can also be used now to FindClosestPointWithinRadius.
Previously it could not be used.