vtkUnstructuredGrid: Improve GetCellNeighbors/IsCellBoundary
The idea behind for GetCellNeighbors' improvement is the fact
that the operation could be done in-place instead of
allocating, de-allocating an extra std::vector

With this change at a specific streamlines experiment GetCellNeighbors
Before changes took 10.958 sec, After changes took  5.542 sec.
The speedup is 10.958 sec / 5.542 sec = 1.982 times better.

GetCellNeighbors is used in many classes:

1) vtkClosestPointStrategy
2) vtkDataSetRegionSurfaceFilter
3) vtkDataSetSurfaceFilter
4) vtkGeometryFilter
5) vtkFillHolesFilter
6) vtkMarkBoundaryFilter
7) vtkPUnstructuredGridGHostCellsGenerator,
8) vtkUnstructuredGridVolumeZSweepMapper,

IsCellBoundary now also has a version that accepts an vtkIdList
to avoid that expensive memory allocation/de-allocation.

IsCellBoundary is used in 2 classes:
1) vtkMarkBoundaryFilter
2) vtkGeometryFilter

All of these classes have greatly benefited from this change.