Raise an error if external data store is needed
For builds from tarballs, if the user turned on VTK_BUILD_TESTING,
cmake used to silently turn it off because the default location for
downloading external data, SOURCE_DIR/.ExternalData, exists in the
VTK git repository, but not in the release tarballs.

After this commit, if SOURCE_DIR/.ExternalData is missing and if
there is no ExternalData or VTKExternalData directory, then cmake
raises a FATAL_ERROR and asks the user to download the VTKData
tarball, which contains the .ExternalData directory.