Cell/Point Locators: UseExistingSearchStructure for all locators
UseExistingSearchStructure, when enabled (default off), it allows the locator to NOT be
built again. This is useful when you have a dataset that either changes because
the FieldData (PointData/CellData) changed or the actual dataset object changed
but it's actually the same geometry (useful when a dataset has timesteps).
When this flag is on you need to use ForceBuildLocator() to rebuild the locator,
if your dataset changes.

To accommodate this functionality now all locators implement the following
functions: BuildLocator/ForceBuildLocator/BuildLocatorInternal.

Also, the LazyEvaluation flag for CellLocators has been deprecated because
it was not used, except in vtkCachingInterpolatedVelocityField, and also it
could lead to thread-safety issues when 2 threads would call e.g. FindCell
and both threads would try to build the locator.