32 bit cell array converter fix
In a few files, the `vtkCellArray` of `vtkUnstructuredGrid` or
`vtkPolyData` was set to 32 bits if there were less than 2^32 points.
Since point ids are signed integers, we can only use 31 bits on a 32
bits array to store point ids. So now the conversion to 32 bits is done
if there are less than 2^31 points.