vtkGeometryFilter: Improve UnstructuredGrid algorithm and fix ghosts
vtkDataSetSurfaceFilter uses a HashTable with a MemoryPool to extract
the surface of an unstructuredGrid. This algorithm has been incorporated
into vtkGeometryFilter. It has also been improved by deleting the duplicate
faces in the hash table (when found), and it has also been multi-threaded.
Also. because this new algorithm was added, the fast mode for uGrids has been removed.
Finally, vtkUnstructuredGridBase is now supported, and also ghost cells are fixed.
RemoveGhostInterfaces provides extra flexibility when using ghost for rendering or algorithms.