Enforce char[256] function param
This is required for istream >> char[]
which now does not work with istream >> char* as of C++20
https://open-std.org/JTC1/SC22/WG21/docs/papers/2018/p0487r1.html

Function array parameters like char[256] will decay to a char* and thus
fail to work with istream, so we need to be strict about the array
parameter type with func(char (&param)[256])

This also means all func callers can only pass a char[256], and there
were a few calls that were using char[1024] (and others with char[256]).
Those have been adjusted to 256, as I could not see any reason to use
the longer buffer - it would have always only read 256 bytes in the
past.