vtkTableBaseClipDataSet: Multithreading
And remove Locator variable since it's not used at all.

It should be noted that when an vtkUnstructuredGrid/vtkPolyData
can not be fully processed, the vtkClipDataSet is used only.
Previously vtkClipDataSet was used only for the cells that couldn't
be clipped fast using the table based algorithm. This created 3
types of issues 1) When e.g. linear and non-linear cells are present
duplicates point merging was not performed, 2) Adds unnecessary
complexity into the code, 3) deteriorates performance if we create
threaded non-linear clipped cells and then merge them because duplicates
removal takes more time, and also we need to use an appendFilter.

For the above reason, I changed the number of expected cells in the
tableBasedClip test.