fix(OpenXR): add support for HP mixed reality controllers
Adds support for HP mixed reality controllers. As these controllers
lack touch input on the joystick this required some changes in
the VR interactor to allow for the joystick or trackpad to initiate
movement without an explicit start event. Most of that code was
already there but some suppression code in OpenXR was preventing
events with a 0.0 x or y from coming through which was wrong.