Null vtkCocoaGLView to vtkCocoaRenderWindow pointer in dtor
vtkCocoaGLView keeps a pointer to the vtkCocoaRenderWindow that created it.
When vtkCocoaRenderWindow (a C++ object) is deallocated, the vtkCocoaGLView
(an Objective-C object) doesn't get notified. Therefore vtkCocoaGLView's
pointer to the vtkCocoaRenderWindow can eventually reference a deallocated
object.

So, now, in the vtkCocoaRenderWindow destructor, check if it was the
vtkCocoaRenderWindow that created the vtkCocoaGLView (something we already
keep track of) and if so null out the relationship.