Use typeid name for templated superclass lookup
To get the superclass for use as tp_base when wrapping VTK classes,
use typeid(type).name() to match what vtkSetGet.h uses for the
ClassName of templated classes.  Without this change, it is not
possible to subclass a templated classe in any modules except for
the one the templated class is defined in.