vtkXMLUnstructuredDataReader fix time series
Time series inside a `.vtu` file could not be read correctly when
updating the timestep. When updating the time step, instead of wiping
the previous output and reading the new time step, the reader was
appending the data. This caused the output to be ill-formed.

Now, on a new time step, the output is cleaned before generating the
current time step.

Adresses paraview/paraview#21445