vtkFeatureEdges: fix behavior with ghost cells.
A regression occured in `vtkFeatureEdges` w.r.t. ghost cells. There are
some instances when one would want that edges be invisible when they are
part of a ghost cell (when rendering for instance), and others where one
would want to keep the interfaces (in some algorithms, such as the ghost
cells generator).

A boolean flag is added: `RemoveGhostInterfaces`, which allows the user
to change the filters' behavior. By default, it is turned on.

Adresses paraview/paraview#21454