vtkPolyDataNormals: Implement Multithreading
Multithread every step of the algorithm that is possible
to be multithread with the existing VTK infrastructure.
Consistency or AutoOrientNormals could not be multithreaded.

vtkSMPTools::SetBackend("SEQUENTIAL"); needed to be added because
the ordering of the addition on normals could vary due to many threads.

Also VTK::CommonDataModelCxx-TestPolyhedronCombinatorialContouring has
been excluded for std-thread because it keeps on timouting due to the
fact that 61,440 really small for loops are launched and the stdthread
backend tries to use all threads.