vtkOutputWindow: make `CurrentMessageType` atomic
The scoping may be done on multiple threads, so it needs to be
atomically set.

Note that there is still an `ABAB` problem here where:

Current state: `X`

- thread A stores the current value `X` and updates the value to `Y`
- thread B stores the current value `Y` and updates it to `Y` itself
- thread A restores it to its idea of the previous value `X`
- thread B restores it to its idea of the previous value `Y`

After state: `Y`

To fix this, the "current message type" should instead be passed around
as needed so that local updates can be made locally instead of having to
consider interactions with other threads.