BUG: Fix QuadricDecimation on Windows
QuadricDecimation was giving incorrect results on Windows when using
Visual Studio 2022 toolset 14.29.30133 or newer. This appears to have been a
compiler optimizer error. The error only happened on Release mode, not
Debug. By using the C++ volatile keyword to inhibit compiler
optimizations around the particular incorrectly optimized away variable,
the correct result was able to be obtained in Release mode as well.

squash