vtkDataReader: fix C++20 compat
This is required for `istream >> char[]` which now does not work with
`istream >> char*` as of C++20

    http://wg21.link/p0487r1

Function array parameters like `char[256]` will decay to a `char*` and
thus fail to work with istream, so we need to be strict about the array
parameter type with `func(char (&param)[256])`.

However, this would technically break compatibility since this
documented-as-internal method is actually `public`, so instead do some
`reinterpret_cast` shenanigans to make the decayed pointer back into a
reference-to-array which contains its size.

Co-authored-by: Paul Harris <harris.pc@gmail.com>
Fixes: #18625