Restoring vtkRedistributeDataSetFilter behavior
!8812 broke a functionality of `vtkRedistributeDataSetFilter`. When
setting `NumberOfPartitions` to some number larger than the number of
input partitions, the filter internally generates a partitioned data set
as a temporary state. When setting `PreservePartitionsInOutput` to
false, the output becomes a `vtkUnstructuredGrid` and we need to merge
the partitions into one partition. Instead of merging, the filter was
just saving the first partition.

The trick to have both this behavior working and not having duplicate
points in the output is to set `MergingPointsOn` in the appender that
merges the partitions into the output.

Fixes #18632