vtkTableBasedClipDataSet: Expedite the clipping for vtkPolyData
It's beneficial to convert a polydata to unstructured grid for clipping because the
GetCellType and GetCellPoints are the most expensive functions used (excluding point/cell
data related functions). The vtkPolyData ones are more expensive than the vtkUnstructuredGrid
ones because they perform a bit operation to get the cell type and then based on that, get
the correct cell array and extract the cell points. This overhead turns out to increase the
execution time by 10%-20%.