Correct HTG locator for non 3d HTGs
Previously the locator would assume that dimensions were only
consecutive, which is not correct (e.g. a 1D HTG can be along the Y
or Z axis, and not only X). The previous algorithm would also
match point that were out of the extent of the dataset. This is
not the case anymore.

Fix tests that would rely on the previous behavior