Add VTK-m override for Clip filter
The test `VTK::FiltersGeneralPython-tableBasedClip` is being disabled
since `vtkmClip` may generate different number of points and cells which
is what is being tested as passing condition.

The slightly different output of `vtkmClip` was causing a slight change
in the result of `TestResampleWithDataSet3`. Using a more accurate cell
locator reduces the difference and the test passes without any baseline
changes.