Disable CVR during interactive rendering
Previously, if CVR was enabled during interactive rendering, the
sampling distance would be reduced extra in order to compensate for
the decreased frame rate.

However, for higher values of Global Illumination Reach in particular,
the reduced sampling distance is not always sufficient to reach the
desired frame rate.

In order to achieve high frame rates during interactive rendering,
disable CVR for now. It is likely possible to look into other options
in the future such as decreasing the Global Illumination Reach.
However, disabling CVR completely during interactive rendering is
likely sufficient for now.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>