Added a new mode to temporal statistics for in situ use.
Added the ability to incrementally update temporal statistics while
timesteps come through a data source such as in situ. In this
situation, it is not possible for the filter to stream time from
its input as only one timestep is available at a time. Instead the
filter works with the current timestep accumulating statistics as
the filter is executed. The filter always outputs the statistics of
all the so far processed timesteps in this mode.