Add a wrapper check function for array refs
In C++, array parameters like "float a[3]" are silently converted
to pointer parameters in the function body.  Some functions work
around this by using pass-by-reference, e.g. "float (&a)[3]",
in which case the parameter 'a' is seen as an array within the
function body instead of as a pointer.

This commit adds a function vtkWrap_IsArrayRef() that the wrappers
can use to check for parameters like "float (&a)[3]".