Add VTK-m override for the Threshold filter
For image data inputs, `vtkmThreshold` generates Hex cells in the output
whereas `vtkThreshold` generates Vox cells. After some discussion, we
have decided to not change vtkm's behaviour to match vtk, as it is
considered as the correct behaviour.

Modified `TestExtractCellsAlongPolyLine` as it was expecting Vox
cells all the time, where as now, it can get Hex cells sometimes.