vtkModule: support skipping installation of modules
Add support for skipping installation of modules. This is useful when
the project is not installing an SDK and does not need to provide
artifacts for the module (e.g., internal tools or static modules that
get bundled into other artifacts).

Note that CMake exports of other targets may contain references to these
targets which can require their presence in the export sets. In this
case, they must be installed regardless of other details (or the project
be restructured to remove their references from the export sets). See
cmake/cmake#23209 for future enhancements to CMake to make this more
possible.

All installation is disabled when this is the case including headers and
license files.